New manager module PART 3: Changed how probing APSIM DLL's works. Instead of ProbeDLL.exe, use UpdateDotNetProxies.exe This now creates a .NET proxy class for all APSIM components. The new manager module (and APSIMUI) can use reflection on these proxy classes to determine variables and events and runtime. This now works for all APSIM and Ausfarm components.
NB: Your manager scripts that used to work with the New Manager Module now won't run without some tweaking.

bugid: 1025

+++ Model/SiloInput/SiloInputComponent.cpp
@@ -64,7 +64,8 @@ SiloInputComponent::SiloInputComponent(void)
 SiloInputComponent::~SiloInputComponent(void)
    {
    data.close();
-   unlink(fileName.c_str());
+   if (fileName != "")
+      unlink(fileName.c_str());
    }
 // ------------------------------------------------------------------
 // INIT 2 - temporary
@@ -86,6 +87,9 @@ void SiloInputComponent::openInputFile(void)
 
    // Now the SILO station number 
    readParameter ("parameters", "station_number", stationNumber, 0, 100000);
+   if (stationNumber == 0)  // This happens during GetDescription call.
+      return;
+   
    baseURL = getProperty("parameters", "url");
    string requestString =
       baseURL +
